# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "POS System API Documentation",
        "description": "Comprehensive Point of Sale System API for multi-tenant restaurant management with authentication, orders, staff, and more.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/admin/restaurants": {
            "get": {
                "tags": [
                    "Admin - Restaurants"
                ],
                "summary": "List all restaurants",
                "description": "Retrieve paginated list of all restaurants (SuperAdmin only)",
                "operationId": "2dc51483b4b903bba04fc02bce0daf39",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by restaurant name, city, or address",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Golden Fork"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by restaurant status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive",
                                "suspended"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page (max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restaurants list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "The Golden Fork"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "golden-fork"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "123 Main Street, Downtown"
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "New York"
                                                    },
                                                    "country": {
                                                        "type": "string",
                                                        "example": "USA"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "+1-555-0101"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "info@goldenfork.com"
                                                    },
                                                    "cuisine_type": {
                                                        "type": "string",
                                                        "example": "American"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "active"
                                                    },
                                                    "subscription_type": {
                                                        "type": "string",
                                                        "example": "premium"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "datetime"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "datetime"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 20
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 4
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - SuperAdmin access required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "JWT Bearer token authentication. Add 'Bearer ' before your token.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}